Big Data Float type

+++ main.py
@@ -107,13 +107,13 @@ class ContinuousScan:
         self.accel_distance = (self.total_distance *
                                self.acceleration_time) / self.total_time
         self.deccel_distance = self.accel_distance
-        return int(self.accel_distance)
+        return float(self.accel_distance)
 
     def calculate_constant_distance(self):
         self.calculate_accel_distance()
         self.constant_distance = self.total_distance - \
             (self.accel_distance + self.deccel_distance)
-        return int(self.constant_distance)
+        return float(self.constant_distance)
 
     def move_epics_motor(self, position):
         # Move the motor to the desired position
